Float maths updates for 2.0.x parity (#11213)

+++ Marlin/printcounter.cpp
@@ -60,7 +60,7 @@ millis_t PrintCounter::deltaDuration() {
   return lastDuration - tmp;
 }
 
-void PrintCounter::incFilamentUsed(double const &amount) {
+void PrintCounter::incFilamentUsed(float const &amount) {
   #if ENABLED(DEBUG_PRINTCOUNTER)
     debug(PSTR("incFilamentUsed"));
   #endif
